fix: remove if exists and fail if path is a file

+++ cmd/soroban-cli/src/commands/contract/bindings/ts.rs
@@ -28,12 +28,20 @@ pub enum Error {
     GenerateTSFromFile(ts::GenerateFromFileError),
     #[error(transparent)]
     Io(#[from] std::io::Error),
+
+    #[error("--root-dir cannot be a file: {0:?}")]
+    IsFile(PathBuf),
 }
 
 impl Cmd {
     pub fn run(&self) -> Result<(), Error> {
         let spec = self.wasm.parse().unwrap().spec;
-        std::fs::remove_dir_all(&self.root_dir)?;
+        if self.root_dir.is_file() {
+            return Err(Error::IsFile(self.root_dir.clone()));
+        }
+        if self.root_dir.exists() {
+            std::fs::remove_dir_all(&self.root_dir)?;
+        }
         std::fs::create_dir(&self.root_dir)?;
         let p: ts::boilerplate::Project = self.root_dir.clone().try_into()?;
         p.init(&self.contract_name, &self.contract_id, &spec)?;
